requirements.txt
psycopg2-binary
pgvector
langchain-openai
langchain-postgres
fastapi
uvicorn
python-dotenv
requests
PyGithub